:zap: refactor: QuestionRenderer, and implement calculateTotalTimeSpent utility
