+++ src/shared/sections/study-session/components/SummaryLabel.tsx
